separate graphql fragements into files
