+++ webapp/graphql/fragments/badges.js
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag'
+
+export const badges = gql`
+  fragment badges on User {
+    badgeTrophiesSelected {
+      id
+      icon
+      description
+    }
+    badgeVerification {
+      id
+      icon
+      description
+    }
+  }
+`
