+++ webapp/graphql/fragments/comment.js
@@ -1,0 +1,17 @@
+import gql from 'graphql-tag'
+
+export const comment = gql`
+  fragment comment on Comment {
+    id
+    createdAt
+    updatedAt
+    disabled
+    deleted
+    content
+    contentExcerpt
+    isPostObservedByMe
+    postObservingUsersCount
+    shoutedByCurrentUser
+    shoutedCount
+  }
+`
