+++ webapp/graphql/fragments/group.js
@@ -1,0 +1,24 @@
+import gql from 'graphql-tag'
+
+export const group = gql`
+  fragment group on Group {
+    id
+    groupName: name
+    slug
+    disabled
+    deleted
+    about
+    description
+    descriptionExcerpt
+    groupType
+    actionRadius
+    categories {
+      id
+      slug
+      name
+      icon
+    }
+    locationName
+    myRole
+  }
+`
