+++ webapp/graphql/fragments/postCounts.js
@@ -1,0 +1,13 @@
+import gql from 'graphql-tag'
+
+export const postCounts = gql`
+  fragment postCounts on Post {
+    commentsCount
+    shoutedCount
+    shoutedByCurrentUser
+    emotionsCount
+    clickedCount
+    viewedTeaserCount
+    viewedTeaserByCurrentUser
+  }
+`
