+++ webapp/graphql/fragments/tagsCategoriesAndPinned.js
@@ -1,0 +1,20 @@
+import gql from 'graphql-tag'
+
+export const tagsCategoriesAndPinned = gql`
+  fragment tagsCategoriesAndPinned on Post {
+    tags {
+      id
+    }
+    categories {
+      id
+      slug
+      name
+      icon
+    }
+    pinnedBy {
+      id
+      name
+      role
+    }
+  }
+`
