+++ webapp/graphql/fragments/userCounts.js
@@ -1,0 +1,12 @@
+import gql from 'graphql-tag'
+
+export const userCounts = gql`
+  fragment userCounts on User {
+    shoutedCount
+    contributionsCount
+    commentedCount
+    followedByCount
+    followingCount
+    followedByCurrentUser
+  }
+`
